Fixes 1
Letters and alphabet length are now variables instead of just copy-paste values. IndexOf now uses the letters variable directly, thereby not requiring a passed value other than character. Lastly, I am freeing up the shifted alphabet as it was pointed out to me that it was a memory leak - more fixes to come